customer-support-agent/requirements.txt
Flask
google-generativeai
grpcio
grpcio-tools